mcp_plugin/requirements.txt
mcp>=1.2.0
nats-py>=2.9.0
